pyproject.toml
[tool.poetry]
name = "mhz_utilities"
version = "0.1.0"
description = ""
authors = ["Zach Mitchell <zmitchell@fastmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
pyserial = "^3.4"
logzero = "^1.5"
click = "^7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
